+++ ReleaseNotes/release_notes.md
@@ -3,7 +3,8 @@
 
   1. add new method `make_soft_constraint2 to create asymmetryc constraint
   1. add test for soft constraints `test_fitting_constraints`
-, hLcpip_Sc, hLcpim_Sc
+  1. add Thiele rational interpolator 
+
 ## Backward incompatible:  
  
 ## Bug fixes: